requirement.txt
streamlit>=1.28.0
pandas>=2.0.0
kagglehub>=0.1.0
python-dotenv>=0.19.0